conversion des Map<String, Integer> en Map<String, Long> pour réduire le code avec les streams

+++ Project02Eclipse/src/com/hemebiotech/analytics/AnalyticsCounter.java
@@ -40,7 +40,7 @@ public class AnalyticsCounter {
 			System.out.println("No symptom was imported.");
 		} else {
 			// count symptoms
-			Map<String, Integer> symptomsMap = (new SymptomCountWithStream()).symptomCountFromList(symptomsList);
+			Map<String, Long> symptomsMap = (new SymptomCountWithStream()).symptomCountFromList(symptomsList);
 			// sort symptoms
 			symptomsMap = (new SymptomSortWithStream()).sortSymptoms(symptomsMap);
 			// export to file
